feat: separate models from AppState
