+++ SNUTT-2022/SNUTT/Models/User.swift
@@ -1,0 +1,14 @@
+//
+//  User.swift
+//  SNUTT
+//
+//  Created by Jinsup Keum on 2022/06/25.
+//
+
+import SwiftUI
+
+class User: ObservableObject {
+    var token: String?
+    var userId: String?
+    var registeredFCMToken: String?
+}
