+++ SNUTT-2022/SNUTT/Views/Scenes/ReviewScene.swift
@@ -13,7 +13,7 @@ struct ReviewScene: View {
     
     var body: some View {
         Button{
-            viewModel.updateTimetable(timeTable: AppState.CurrentTimetable())
+            viewModel.updateTimetable(timeTable: Timetable(lectures: []))
         } label: {
             Text("Change current timetable")
         }
